Splits getalvelden.class.php
