+++ lib/view/formulier/getalvelden/TelefoonField.class.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * TelefoonField.class.php
+ *
+ * @author Jan Pieter Waagmeester <jieter@jpwaag.com>
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * is valid als er een enigszins op een telefoonnummer lijkende string wordt
+ * ingegeven.
+ */
+class TelefoonField extends TextField {
+
+	public function validate() {
+		if (!parent::validate()) {
+			return false;
+		}
+		// parent checks not null
+		if ($this->value == '') {
+			return true;
+		}
+		if (!preg_match('/^([\d\+\-]{10,20})$/', $this->value)) {
+			$this->error = 'Geen geldig telefoonnummer.';
+		}
+		return $this->error === '';
+	}
+
+}
